[feature/generate_file_by_action]feat: make file by action
