+++ packages/backend/src/plugins/scaffolder/actions/create-file/index.ts
@@ -1,0 +1,1 @@
+export { createNewFileAction } from "./dependcy";
